Refactorisation of the code of AddFlight in FlightForm and add of ModifyFlight
